finishing up migrate data for local db

+++ Resources/model/chatHistory.js
@@ -35,47 +35,56 @@ exports.getChatHistory = function(_matchId) {
 };
 
 exports.migrateData = function() {
+	Ti.API.info('migrating chatHistory data..');
 	//drop table then recreate it
 	var db = Ti.Database.open(Ti.App.DATABASE_NAME); 
 	var chatMessages = [];
-	var result = db.execute("SELECT * FROM ChatHistory");
-	var numChatsInDb = 0;
-	while(result.isValidRow()) {
-		var chatObj = {}; 
-		if(result.fieldByName('MatchId') !== undefined && result.fieldByName('MatchId') !== null) {
-			chatObj.matchId = result.fieldByName('MatchId'); 
-			chatObj.userId = result.fieldByName('UserId'); 
-			chatObj.targetedUserId = result.fieldByName('TargetedUserId'); 
-			chatObj.senderId = result.fieldByName('SenderId'); 
-			chatObj.receiverId = result.fieldByName('ReceiverId'); 
-			chatObj.message = result.fieldByName('Message'); 
-			chatObj.time = result.fieldByName('Time'); 
+	var resultCheckingColumn = null;
+	var hasMatchIdColumn = true;
+	try {
+		resultCheckingColumn = db.execute("SELECT MatchId FROM ChatHistory LIMIT 0,1");		
+	} catch(e) {
+		hasMatchIdColumn = false;
+	}
+	if(resultCheckingColumn !== null)
+		resultCheckingColumn.close();
+	
+	if(hasMatchIdColumn) {	
+		var resultChatHistory = db.execute("SELECT * FROM ChatHistory");
+		var numChatsInDb = 0;
+		while(resultChatHistory.isValidRow()) {
+			var chatObj = {}; 
+			chatObj.matchId = resultChatHistory.fieldByName('MatchId'); 
+			chatObj.userId = resultChatHistory.fieldByName('UserId'); 
+			chatObj.targetedUserId = resultChatHistory.fieldByName('TargetedUserId'); 
+			chatObj.senderId = resultChatHistory.fieldByName('SenderId'); 
+			chatObj.receiverId = resultChatHistory.fieldByName('ReceiverId'); 
+			chatObj.message = resultChatHistory.fieldByName('Message'); 
+			chatObj.time = resultChatHistory.fieldByName('Time'); 
 			chatMessages.push(chatObj);
-		} //else //ignore msg that does not have matchId
-		result.next();
-		numChatsInDb++;
-	}
-	//alert('numChatsInDb before migrate: '+numChatsInDb);
-	
-	db.execute("DROP TABLE IF EXISTS ChatHistory");
-	
-	db.execute('CREATE TABLE ChatHistory(Id INTEGER PRIMARY KEY, MatchId INTEGER, UserId INTEGER, TargetedUserId INTEGER, SenderId INTEGER, ReceiverId INTEGER, Message TEXT, Time TEXT);');
-	var numChatsMigrated = 0;
-	for(var i = 0; i < chatMessages.length; i++) {
+			resultChatHistory.next();
+			numChatsInDb++;
+		}
+		//alert('numChatsInDb before migrate: '+numChatsInDb);
+		resultChatHistory.close();
 		
-		var chatObj = chatMessages[i]; 
-		var matchId = chatObj.matchId;
-		var userId = chatObj.userId;
-		var targetedUserId = chatObj.targetedUserId; 
-		var senderId = chatObj.senderId;
-		var receiverId = chatObj.receiverId;
-		var message = chatObj.message;
-		var sentTime = chatObj.time;
-		db.execute("INSERT INTO ChatHistory(Id, MatchId, UserId, TargetedUserId, SenderId, ReceiverId, Message, Time) VALUES(NULL,?,?,?,?,?,?,?)", matchId, userId, targetedUserId, senderId, receiverId, message, sentTime);	
-		numChatsMigrated++;
-	}
-	//alert('numChatsMigrated: '+numChatsMigrated);
-	
-	result.close();
+		db.execute("DROP TABLE IF EXISTS ChatHistory");
+		db.execute('CREATE TABLE ChatHistory(Id INTEGER PRIMARY KEY, MatchId INTEGER, UserId INTEGER, TargetedUserId INTEGER, SenderId INTEGER, ReceiverId INTEGER, Message TEXT, Time TEXT);');
+		var numChatsMigrated = 0;
+		for(var i = 0; i < chatMessages.length; i++) {
+			
+			var chatObj = chatMessages[i]; 
+			var matchId = chatObj.matchId;
+			var userId = chatObj.userId;
+			var targetedUserId = chatObj.targetedUserId; 
+			var senderId = chatObj.senderId;
+			var receiverId = chatObj.receiverId;
+			var message = chatObj.message;
+			var sentTime = chatObj.time;
+			db.execute("INSERT INTO ChatHistory(Id, MatchId, UserId, TargetedUserId, SenderId, ReceiverId, Message, Time) VALUES(NULL,?,?,?,?,?,?,?)", matchId, userId, targetedUserId, senderId, receiverId, message, sentTime);	
+			numChatsMigrated++;
+		}
+		//alert('numChatsMigrated: '+numChatsMigrated);
+	}	
 	db.close();
 };
